add prefix label pass for lexer
