+++ src/assembler/mod.rs
@@ -1,11 +1,25 @@
-use crate::defs::LC3Word;
+use crate::defs::{LC3Word, Op, PseudoOp, RegAddr};
 
 pub mod tokenizer;
+pub mod lexer;
 
 pub struct MaybeUnresolvedInstr {
     value: LC3Word,
     ///Label, Start offset, End offset
     bindings: Option<(String, u8, u8)>,
+}
+
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub enum Token {
+    INSTR(Op),
+    REGISTER(RegAddr),
+    META(PseudoOp),
+    STRING(String),
+    NUM(LC3Word),
+    COMMENT(String),
+    QUOTES,
+    SEMICOLON,
+    COMMA,
 }
 
 pub fn translate_line(line: &str) -> MaybeUnresolvedInstr {
